- Fixed several compilatoin warnings from wrong formatting string used with numeric types [TESTS] - full build with gcc

+++ library/grt/src/grtpp_grt.cpp
@@ -1044,7 +1044,7 @@ bool GRT::handle_message(const Message &msg, void *sender)
         return true;
     }
   }
-  log_error("Unhandled message (%lu): %s\n", _message_slot_stack.size(), msg.format().c_str());
+  log_error("Unhandled message (%lu): %s\n", (unsigned long)_message_slot_stack.size(), msg.format().c_str());
   return false;
 }
